+++ mech/mech_instance.py
@@ -125,6 +125,12 @@ class MechInstance():
                 self.vbox = None
                 self.created = False
 
+        self.no_nat = False
+        self.remove_vagrant = False
+        self.gui = False
+        self.disable_provisioning = False
+        self.disable_shared_folders = False
+
         # If vmx exists, then the VM has already been created.
         # See if we need to switch to preshared key authentication
         # for interactions with this guest.
@@ -192,13 +198,16 @@ class MechInstance():
     def __repr__(self):
         """Return a representation of a Mech instance."""
         sep = '\n'
-        return ('name:{name}{sep}created:{created}{sep}box:{box}{sep}'
-                'box_version:{box_version}{sep}'
-                'url:{url}{sep}box_file:{box_file}{sep}provision:{provision}{sep}'
-                'vmx:{vmx}{sep}user:{user}{sep}'
-                'password:{password}{sep}enable_ip_lookup:{enable_ip_lookup}'
+        return ('name:{name}{sep}created:{created}{sep}box:{box}'
+                '{sep}box_version:{box_version}'
+                '{sep}url:{url}{sep}box_file:{box_file}{sep}provision:{provision}'
+                '{sep}vmx:{vmx}{sep}user:{user}'
+                '{sep}password:{password}{sep}enable_ip_lookup:{enable_ip_lookup}'
                 '{sep}config:{config}{sep}shared_folders:{shared_folders}'
-                '{sep}auth:{auth}{sep}use_psk:{use_psk}{sep}provider:{provider}'.
+                '{sep}auth:{auth}{sep}use_psk:{use_psk}{sep}provider:{provider}'
+                '{sep}no_nat:{no_nat}{sep}remove_vagrant:{remove_vagrant}'
+                '{sep}gui:{gui}{sep}disable_provisioning:{disable_provisioning}'
+                '{sep}disable_shared_folders:{disable_shared_folders}'.
                 format(name=self.name, created=self.created,
                        box=self.box, box_version=self.box_version,
                        url=self.url, box_file=self.box_file,
@@ -208,7 +217,11 @@ class MechInstance():
                        config=self.config,
                        shared_folders=self.shared_folders,
                        auth=self.auth, use_psk=self.use_psk,
-                       provider=self.provider, sep=sep))
+                       provider=self.provider, no_nat=self.no_nat,
+                       remove_vagrant=self.remove_vagrant, gui=self.gui,
+                       disable_provisioning=self.disable_provisioning,
+                       disable_shared_folders=self.disable_shared_folders,
+                       sep=sep))
 
     def config_ssh(self):
         """Configure ssh to work. If needed, create an insecure private key file for ssh/scp."""